+++ custom_components/proxmox_sensors/sensor/sensor_last_action.py
@@ -13,13 +13,13 @@ class PBSLastActionSensor(CoordinatorEntity, SensorEntity):
 
     _attr_has_entity_name = True
 
-    def __init__(self, coordinator, datastore):
+    def __init__(self, coordinator, server_id, datastore):
         super().__init__(coordinator)
 
         self._datastore = datastore
 
         self._attr_translation_key = "last_action"
-        self._attr_unique_id = f"{datastore.lower()}_last_action"
+        self._attr_unique_id = f"pbs_{server_id}_{datastore}_last_action"
         self._attr_should_poll = False
 
         self._state = STATE_UNKNOWN